app/requirements.txt
streamlit
torch
torchvision
torch_utils
pillow
numpy